+++ src/lang-xpath/test/uk/co/reecedunn/intellij/plugin/xpath/tests/parser/PluginParserTest.kt
@@ -38,8 +38,8 @@ private class PluginParserTest : ParserTestCase() {
     }
 
     @Nested
-    @DisplayName("XQuery IntelliJ Plugin EBNF (22) UnionType")
-    internal inner class UnionType {
+    @DisplayName("XQuery IntelliJ Plugin EBNF (22) UnionType ; XPath 3.1 EBNF (52) ItemType")
+    internal inner class UnionType_ItemType {
         @Test
         @DisplayName("NCName")
         fun ncname() {
@@ -120,14 +120,34 @@ private class PluginParserTest : ParserTestCase() {
             val actual = parseResource("tests/parser/xpath-ng/xpath/proposal-6/UnionType_Multiple.xq")
             assertThat(prettyPrintASTNode(actual), `is`(expected))
         }
+    }
 
+    @Nested
+    @DisplayName("XQuery IntelliJ Plugin EBNF (22) UnionType ; XPath 3.1 EBNF (77) SingleType")
+    internal inner class UnionType_SingleType {
         @Test
-        @DisplayName("union in TypedMapTest")
-        fun testUnionType_InTypedMapTest() {
-            val expected = loadResource("tests/parser/xpath-ng/xpath/proposal-6/UnionType_InTypedMapTest.txt")
-            val actual = parseResource("tests/parser/xpath-ng/xpath/proposal-6/UnionType_InTypedMapTest.xq")
+        @DisplayName("single type")
+        fun singleType() {
+            val expected = loadResource("tests/parser/xpath-ng/xpath/proposal-6/UnionType_InSingleType.txt")
+            val actual = parseResource("tests/parser/xpath-ng/xpath/proposal-6/UnionType_InSingleType.xq")
             assertThat(prettyPrintASTNode(actual), `is`(expected))
         }
+
+        @Test
+        @DisplayName("optional")
+        fun optional() {
+            val expected = loadResource("tests/parser/xpath-ng/xpath/proposal-6/UnionType_InSingleType_Optional.txt")
+            val actual = parseResource("tests/parser/xpath-ng/xpath/proposal-6/UnionType_InSingleType_Optional.xq")
+            assertThat(prettyPrintASTNode(actual), `is`(expected))
+        }
+    }
+
+    @Test
+    @DisplayName("XQuery IntelliJ Plugin EBNF (22) UnionType ; XPath 3.1 EBNF (107) TypedMapTest")
+    fun unionType_typedMapTest() {
+        val expected = loadResource("tests/parser/xpath-ng/xpath/proposal-6/UnionType_InTypedMapTest.txt")
+        val actual = parseResource("tests/parser/xpath-ng/xpath/proposal-6/UnionType_InTypedMapTest.xq")
+        assertThat(prettyPrintASTNode(actual), `is`(expected))
     }
 
     @Nested